+++ packages/button-react/documentation/Example.tsx
@@ -5,6 +5,7 @@ import "@fremtind/jkl-toggle-switch/toggle-switch.css";
 import "@fremtind/jkl-button/button.css";
 import "@fremtind/jkl-core/core.css";
 import "./index.scss";
+import classNames from "classnames";
 
 function onClick() {
     console.log("Hello!");
@@ -16,17 +17,19 @@ const Example = () => {
     const [isInverted, setIsInverted] = useState(false);
     const toggleInverted = () => setIsInverted(!isInverted);
 
-    const exampleClassName = "buttons-example jkl-spacing--top-3 jkl-spacing--bottom-3".concat(
-        isInverted ? " buttons-example--inverted" : "",
-    );
+    const exampleClassName = classNames("buttons-example jkl-spacing--top-3 jkl-spacing--bottom-3", {
+        "buttons-example--inverted": isInverted,
+    });
 
     return (
         <section className={exampleClassName}>
             <aside className="buttons-example__controls">
-                <ToggleSwitch onChange={toggleCompact} className="jkl-spacing--bottom-1">
+                <ToggleSwitch pressed={isCompact} onClick={toggleCompact} className="jkl-spacing--bottom-1">
                     Kompakt variant
                 </ToggleSwitch>
-                <ToggleSwitch onChange={toggleInverted}>Inverterte farger</ToggleSwitch>
+                <ToggleSwitch pressed={isInverted} onClick={toggleInverted}>
+                    Inverterte farger
+                </ToggleSwitch>
             </aside>
             <PrimaryButton
                 forceCompact={isCompact}