+++ packages/checkbox-react/src/Checkbox.tsx
@@ -1,4 +1,5 @@
 import React, { ReactNode, ChangeEvent } from "react";
+import classNames from "classnames";
 
 interface Props {
     children: ReactNode;
@@ -23,13 +24,14 @@ export function Checkbox({
     inline = false,
     forceCompact,
 }: Props) {
-    const classNames = "jkl-checkbox"
-        .concat(forceCompact ? " jkl-checkbox--compact" : "")
-        .concat(inline ? " jkl-checkbox--inline" : "")
-        .concat(invalid ? " jkl-checkbox--error" : "")
-        .concat(className ? ` ${className}` : "");
+    const componentClassNames = classNames("jkl-checkbox", className, {
+        "jkl-checkbox--compact": forceCompact,
+        "jkl-checkbox--inline": inline,
+        "jkl-checkbox--error": invalid,
+    });
+
     return (
-        <label className={classNames}>
+        <label className={componentClassNames}>
             <input
                 className="jkl-checkbox__input"
                 data-testid="jkl-checkbox-input"