+++ packages/datepicker-react/src/DatePicker.tsx
@@ -1,12 +1,21 @@
-import React, { ChangeEvent, useRef, useState, useEffect } from "react";
+import React, { ChangeEvent, useState, useEffect, useRef } from "react";
+import nanoid from "nanoid";
 import { Label, SupportLabel } from "@fremtind/jkl-typography-react";
-import { useClickOutside, useFocusOutside, useKeyListener } from "@fremtind/jkl-react-hooks";
 import { LabelVariant } from "@fremtind/jkl-core";
+import { CalendarIcon } from "./CalendarIcon";
 // @ts-ignore
 import CoreDatepicker from "@nrk/core-datepicker/jsx";
+// @ts-ignore
+import CoreToggle from "@nrk/core-toggle/jsx";
+import classNames from "classnames";
 
 interface ChangeDate {
     date: Date;
+}
+
+interface CoreToggleSelectEvent {
+    detail: { textContent: string; value: string; classList: DOMTokenList };
+    target: { hidden: boolean; button: HTMLButtonElement; value: { textContent: string } };
 }
 
 interface Props {
@@ -16,6 +25,7 @@ interface Props {
     placeholder?: string;
     months?: string[];
     days?: string[];
+    calendarButtonTitle?: string;
     initialDate?: Date;
     onChange?: (date: Date) => void;
     extended?: boolean;
@@ -57,6 +67,7 @@ export function DatePicker({
     placeholder = "dd.mm.åååå",
     months,
     days,
+    calendarButtonTitle = "Vis kalender",
     initialDate,
     onChange,
     extended = false,
@@ -69,17 +80,21 @@ export function DatePicker({
     disableBeforeDate,
     disableAfterDate,
 }: Props) {
+    const { current: uuid } = useRef(`jkl-datepicker-${nanoid(8)}`);
     const [date, setDate] = useState(initialDate);
     const [datepickerHidden, setDatepickerHidden] = useState(!initialShow);
     const [dateString, setDateString] = useState(initialDate ? formatDate(initialDate) : "");
-    const componentClassName = "jkl-datepicker".concat(
-        extended ? " jkl-datepicker--extended" : "",
-        !datepickerHidden ? " jkl-datepicker--open" : "",
-        className ? ` ${className}` : "",
+    const componentClassName = classNames(
+        "jkl-datepicker",
+        {
+            "jkl-datepicker--extended": extended,
+            "jkl-datepicker--open": !datepickerHidden,
+        },
+        className,
     );
-    const inputClassName = "jkl-text-field jkl-datepicker__input".concat(
-        forceCompact ? ` jkl-text-field--compact` : "",
-    );
+    const inputClassName = classNames("jkl-text-field jkl-text-field--action jkl-datepicker__input", {
+        "jkl-text-field--compact": forceCompact,
+    });
 
     useEffect(() => {
         disableAfterDate && disableAfterDate.setHours(23, 59, 59, 999);
@@ -93,19 +108,12 @@ export function DatePicker({
         }
     }, [disableBeforeDate, disableAfterDate]);
 
-    const openDatepicker = (e: React.FocusEvent<HTMLInputElement>) => {
-        // Workaround for loosing focus when opening in chrome:
-        // https://github.com/nrkno/core-components/issues/322
-        e.persist();
-        setTimeout(() => e.target.focus(), 0);
-        setDatepickerHidden(false);
+    const toggleDatepicker = () => setDatepickerHidden(!datepickerHidden);
+    const handleCalendarClick = (e: CoreToggleSelectEvent) => {
+        if (!e.detail.classList.contains("jkl-datepicker__month-button")) {
+            e.target.hidden = true;
+        }
     };
-    const closeDatepicker = () => setDatepickerHidden(true);
-
-    const datepickerRef = useRef<HTMLDivElement>(null);
-    useClickOutside(datepickerRef, closeDatepicker);
-    useFocusOutside(datepickerRef, closeDatepicker);
-    useKeyListener(datepickerRef, ["Enter", "Escape"], closeDatepicker);
 
     function onInputChange(event: ChangeEvent<HTMLInputElement>) {
         const newDateString = event.target.value;
@@ -133,7 +141,6 @@ export function DatePicker({
 
     function onClickCalendarDay(event: ChangeEvent<ChangeDate>) {
         const newDate = event.target.date;
-        setDatepickerHidden(true);
 
         if (dateHasChanged(date, newDate)) {
             setDateString(formatDate(newDate));
@@ -150,25 +157,38 @@ export function DatePicker({
     }
 
     return (
-        <div className={componentClassName} ref={datepickerRef}>
-            <div className="jkl-datepicker__outer-wrapper">
-                <label className={inputClassName}>
-                    <Label variant={variant} forceCompact={forceCompact}>
-                        {label}
-                    </Label>
-                    <input
-                        placeholder={placeholder}
-                        type="text"
-                        aria-invalid={!!errorLabel}
-                        className={`jkl-text-field__input`}
-                        data-testid="jkl-datepicker__input"
-                        value={dateString}
-                        onChange={onInputChange}
-                        onFocus={openDatepicker}
-                    />
-                </label>
-
-                <div hidden={datepickerHidden}>
+        <div className={componentClassName}>
+            <Label htmlFor={uuid} standAlone variant={variant} forceCompact={forceCompact}>
+                {label}
+            </Label>
+            <div className={inputClassName}>
+                <input
+                    id={uuid}
+                    placeholder={placeholder}
+                    type="text"
+                    aria-invalid={!!errorLabel}
+                    className="jkl-text-field__input"
+                    data-testid="jkl-datepicker__input"
+                    value={dateString}
+                    onChange={onInputChange}
+                />
+                <button
+                    type="button"
+                    title={calendarButtonTitle}
+                    data-testid="jkl-datepicker__calendar-button"
+                    className="jkl-text-field__action-button"
+                >
+                    <CalendarIcon className="jkl-text-field__action-icon" />
+                    <span data-testid="jkl-datepicker__calendar-button-text" className="jkl-sr-only">
+                        {calendarButtonTitle}
+                    </span>
+                </button>
+                <CoreToggle
+                    popup
+                    hidden={datepickerHidden}
+                    onToggle={toggleDatepicker}
+                    onToggleSelect={handleCalendarClick}
+                >
                     <CoreDatepicker
                         timestamp={date ? date.getTime() : undefined}
                         months={months}
@@ -213,7 +233,7 @@ export function DatePicker({
                         )}
                         <table data-testid="jkl-datepicker-calendar" />
                     </CoreDatepicker>
-                </div>
+                </CoreToggle>
             </div>
             <SupportLabel errorLabel={errorLabel} helpLabel={helpLabel} forceCompact={forceCompact} />
         </div>