+++ packages/logo-react/src/index.ts
@@ -1,17 +1,7 @@
-import { LogoSymbol } from "./LogoSymbol";
-import { LogoMain } from "./LogoMain";
-import { LogoAnimated } from "./LogoAnimated";
+import { Logo } from "./Logo";
 import { LogoStampTechnology } from "./stamp/LogoStampTechnology";
 import { LogoStampDistributor } from "./stamp/LogoStampDistributor";
 import { LogoStampInnovation } from "./stamp/LogoStampInnovation";
 import { LogoStampInsurance } from "./stamp/LogoStampInsurance";
 
-export {
-    LogoSymbol,
-    LogoMain,
-    LogoAnimated,
-    LogoStampTechnology,
-    LogoStampDistributor,
-    LogoStampInnovation,
-    LogoStampInsurance,
-};
+export { Logo, LogoStampTechnology, LogoStampDistributor, LogoStampInnovation, LogoStampInsurance };