+++ packages/select-react/src/Select.test.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 import { cleanup, render } from "@testing-library/react";
 import { Select } from ".";
+import { axe } from "jest-axe";
+
+afterEach(cleanup);
 
 describe("Select", () => {
-    afterEach(cleanup);
-
     it("should render correct amount of options", () => {
         const { getAllByTestId } = render(<Select items={["drop", "it", "like", "its", "hot"]} label="Snoop" />);
 
@@ -55,5 +56,31 @@ describe("Select", () => {
         );
 
         expect(getByTestId("jkl-select__value").innerHTML).toBe("Fin lesbar tekst");
+    });
+});
+
+describe("a11y", () => {
+    test("select should be a11y compliant", async () => {
+        const { container } = render(<Select label="Select" items={["1", "2"]} value="1" helpLabel="Velg en av to" />);
+        const results = await axe(container, {
+            rules: {
+                "aria-input-field-name": { enabled: false },
+            },
+        });
+
+        expect(results).toHaveNoViolations();
+    });
+
+    test("compact select should be a11y compliant", async () => {
+        const { container } = render(
+            <Select forceCompact label="Select" items={["1", "2"]} value="1" helpLabel="Velg en av to" />,
+        );
+        const results = await axe(container, {
+            rules: {
+                "aria-input-field-name": { enabled: false },
+            },
+        });
+
+        expect(results).toHaveNoViolations();
     });
 });